Remove all static ui data and fix memory leak on exit
